refactor: extract Invokable interface
