+++ basyx.submodelservice/basyx.submodelservice-core/src/main/java/org/eclipse/digitaltwin/basyx/operation/OperationRequestExecutor.java
@@ -70,9 +70,4 @@ public class OperationRequestExecutor {
         return new DefaultOperationResult.Builder().success(true).outputArguments(outputVars).inoutputArguments(inoutputVars).build();
     }
 
-    @FunctionalInterface
-    public interface Invokable {
-        OperationVariable[] invoke(OperationVariable[] arguments);
-    }
-
 }
